+++ xbmc/addons/AddonStatusHandler.cpp
@@ -103,7 +103,7 @@ void CAddonStatusHandler::Process()
       pDialog->SetLine(2, 24073);
 
       //send message and wait for user input
-      ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_YES_NO, (unsigned int)g_windowManager.GetActiveWindow()};
+      ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_YES_NO, g_windowManager.GetActiveWindow()};
       CApplicationMessenger::Get().SendMessage(tMsg, true);
 
       if (pDialog->IsConfirmed())
@@ -120,7 +120,7 @@ void CAddonStatusHandler::Process()
     pDialog->SetLine(1, 24074);
 
     //send message and wait for user input
-    ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_OK, (unsigned int)g_windowManager.GetActiveWindow()};
+    ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_OK, g_windowManager.GetActiveWindow()};
     CApplicationMessenger::Get().SendMessage(tMsg, true);
 
     CAddonMgr::Get().GetCallbackForType(m_addon->Type())->RequestRestart(m_addon, true);
@@ -137,7 +137,7 @@ void CAddonStatusHandler::Process()
     pDialogYesNo->SetLine(3, m_message);
 
     //send message and wait for user input
-    ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_YES_NO, (unsigned int)g_windowManager.GetActiveWindow()};
+    ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_YES_NO, g_windowManager.GetActiveWindow()};
     CApplicationMessenger::Get().SendMessage(tMsg, true);
 
     if (!pDialogYesNo->IsConfirmed()) return;
@@ -165,7 +165,7 @@ void CAddonStatusHandler::Process()
     pDialog->SetLine(3, m_message);
 
     //send message and wait for user input
-    ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_OK, (unsigned int)g_windowManager.GetActiveWindow()};
+    ThreadMessage tMsg = {TMSG_DIALOG_DOMODAL, WINDOW_DIALOG_OK, g_windowManager.GetActiveWindow()};
     CApplicationMessenger::Get().SendMessage(tMsg, true);
   }
 }
